Add const.py, move default creation options into it. Also definitions of INTERSECTION, UNION, etc.
