+++ rios/imagewriter.py
@@ -31,75 +31,7 @@ from . import rioserrors
 from . import rat
 from . import calcstats
 from . import fileinfo
-
-
-def setDefaultDriver():
-    """
-    Sets some default values into global variables, defining
-    what defaults we should use for GDAL driver. On any given
-    output file these can be over-ridden, and can be over-ridden globally
-    using the environment variables
-
-        * $RIOS_DFLT_DRIVER
-        * $RIOS_DFLT_DRIVEROPTIONS
-        * $RIOS_DFLT_CREOPT_<drivername>
-    
-    If RIOS_DFLT_DRIVER is set, then it should be a gdal short driver name. 
-    If RIOS_DFLT_DRIVEROPTIONS is set, it should be a space-separated list
-    of driver creation options, e.g. "COMPRESS=LZW TILED=YES", and should
-    be appropriate for the selected GDAL driver. This can also be 'None'
-    in which case an empty list of creation options is passed to the driver.
-    
-    The same rules apply to the driver-specific creation options given
-    using $RIOS_DFLT_CREOPT_<driver>. These options are a later paradigm, and 
-    are intended to supercede the previous generic driver defaults. 
-    
-    If not otherwise supplied, the default is to use the HFA driver, with compression. 
-    
-    The code here is more complex than desirable, because it copes with legacy behaviour
-    in the absence of the environment variables, and in the absence of the driver-specific
-    option variables. 
-        
-    """
-    global DEFAULTDRIVERNAME, DEFAULTCREATIONOPTIONS
-    DEFAULTDRIVERNAME = os.getenv('RIOS_DFLT_DRIVER', default='HFA')
-    creationOptionsStr = os.getenv('RIOS_DFLT_DRIVEROPTIONS')
-    if creationOptionsStr is not None:
-        DEFAULTCREATIONOPTIONS = creationOptionsStr.split()
-    else:
-        # To cope with the old behaviour, set something sensible for HFA, but not
-        # otherwise.
-        # The IGNOREUTM=YES is there to switch off a minor kludge in GDAL's
-        # HFA driver. By default, it will check any Transverse Mercator
-        # projection, and if its parameters match a standard UTM zone, it
-        # re-states the projection as literal UTM. This was originally because
-        # Imagine was not good at matching equivalent projections. This is no
-        # longer true, and we choose to disable that behaviour by default.
-        if DEFAULTDRIVERNAME == "HFA":
-            DEFAULTCREATIONOPTIONS = ['COMPRESSED=YES', 'IGNOREUTM=YES']
-        else:
-            DEFAULTCREATIONOPTIONS = []
-    
-    # In the new paradigm, default creation options are specific to each driver, and
-    # are loaded into a dictionary
-    global dfltDriverOptions
-    dfltDriverOptions = {}
-    # Start with the old generic default options, applied to the default driver
-    dfltDriverOptions[DEFAULTDRIVERNAME] = DEFAULTCREATIONOPTIONS
-    # Load some others which we wish to have as defaults, even if not set by the environment
-    dfltDriverOptions['GTiff'] = ['TILED=YES', 'COMPRESS=DEFLATE',
-        'INTERLEAVE=BAND', 'BIGTIFF=IF_SAFER']
-    # Now load any which are specified by environment variables, of the
-    # form RIOS_DFLT_CREOPT_<drivername>
-    driverOptVarPrefix = 'RIOS_DFLT_CREOPT_'
-    for varname in os.environ:
-        if varname.startswith(driverOptVarPrefix):
-            drvrName = varname[len(driverOptVarPrefix):]
-            optionsStr = os.getenv(varname)
-            dfltDriverOptions[drvrName] = optionsStr.split()
-
-
-setDefaultDriver()
+from . import const
 
 
 def writeBlock(gdalOutObjCache, blockDefn, outfiles, outputs, controls,
@@ -151,7 +83,7 @@ def openOutfile(symbolicName, filename, controls, arr, workinggrid):
     creationoptions = controls.getOptionForImagename('creationoptions',
         symbolicName)
     if creationoptions is None:
-        creationoptions = dfltDriverOptions.get(driverName, [])
+        creationoptions = const.dfltDriverOptions.get(driverName, [])
     doubleCheckCreationOptions(driverName, creationoptions, controls,
         workinggrid)
